Added codestyle and license header. Updated JNA to newest stable release.
