+++ nativec-jna-impl/pom.xml
@@ -1,0 +1,83 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <groupId>se.jguru.codestyle.poms.java</groupId>
+    <artifactId>jguru-codestyle-java-parent</artifactId>
+    <version>1.2.0</version>
+    <relativePath />
+  </parent>
+
+  <groupId>de.intarsys.opensource.nativec.impl.jna</groupId>
+  <artifactId>nativec-jna-impl</artifactId>
+  <version>1.0.0-SNAPSHOT</version>
+  <packaging>jar</packaging>
+
+  <properties>
+    <license.id>intarsys</license.id>
+  </properties>
+
+  <inceptionYear>2013</inceptionYear>
+
+  <organization>
+    <name>Intarsys Consulting GmbH</name>
+    <url>https://www.intarsys.de</url>
+  </organization>
+
+  <licenses>
+    <license>
+      <name>Intarsys License</name>
+      <url>https://www.intarsys.de/licenses/intarsysSourceLicense.txt</url>
+      <distribution>repo</distribution>
+      <comments>A business-friendly OSS license</comments>
+    </license>
+  </licenses>
+
+  <dependencies>
+    <dependency>
+      <groupId>de.intarsys.opensource</groupId>
+      <artifactId>native-c-api</artifactId>
+      <version>1.0.0-SNAPSHOT</version>
+    </dependency>
+
+    <dependency>
+      <groupId>net.java.dev.jna</groupId>
+      <artifactId>jna</artifactId>
+      <version>5.11.0</version>
+    </dependency>
+
+    <dependency>
+      <groupId>de.intarsys.opensource.nativec.codestyle</groupId>
+      <artifactId>nativec-codestyle</artifactId>
+      <version>1.0.0-SNAPSHOT</version>
+      <scope>compile</scope>
+    </dependency>
+
+  </dependencies>
+
+  <build>
+    <plugins>
+
+      <!--
+          Ensure that the software license is properly defined within the project.
+        -->
+      <plugin>
+        <groupId>org.codehaus.mojo</groupId>
+        <artifactId>license-maven-plugin</artifactId>
+        <version>2.0.0</version>
+
+        <dependencies>
+          <dependency>
+            <groupId>de.intarsys.opensource.nativec.codestyle</groupId>
+            <artifactId>nativec-codestyle</artifactId>
+            <version>1.0.0-SNAPSHOT</version>
+          </dependency>
+        </dependencies>
+      </plugin>
+    </plugins>
+  </build>
+
+</project>
