+++ nativec-jna-impl/src/main/resources/META-INF/services/de.intarsys.nativec.api.INativeInterface
